Correction of bug.

+++ Courseworks/201516/One/Cases.java
@@ -43,7 +43,7 @@ class Cases {
     {
       long sum = 0;
       for (int i = 1, mult=1; i < nargs; ++i, mult*=-1)
-        sum += (mult * Integer.parseInt(args[i]));
+        sum += mult * (long)Integer.parseInt(args[i]);
       final int s0 = Integer.parseInt(args[0]);
       if (s0 % 2 == 0) {sum *= -1; sum += s0;} else sum -= s0;
       System.out.println(sum);
